Chore(set up Postgress): Set up Postgress database
- Download Postgress in the system
- Install the database
- Versioning the project
- removes coveralls scripts and module
- edit test files in order to pass
- make create table in  database function

[finishes #169530831]
