+++ package.json
@@ -10,8 +10,8 @@
   "scripts": {
     "start": "babel-node server.js",
     "dev": "nodemon --exec babel-node server.js",
-    "test": "nyc --reporter=html --reporter=text mocha --require @babel/register --require @babel/polyfill --timeout 500 server/tests/ --exit",
-    "coveralls": "nyc report --reporter=text-lcov | coveralls"
+    "test": "nyc --reporter=html --reporter=text mocha --require @babel/register --require @babel/polyfill --timeout 500 server/v2/tests --exit",
+    "createTables": "NODE_ENV=development babel-node ./server/v2/models/tableCreator.js"
   },
   "repository": {
     "type": "git",
@@ -42,14 +42,14 @@
     "jsonwebtoken": "^8.5.1",
     "morgan": "^1.9.1",
     "nodemon": "^1.19.1",
-    "nyc": "^14.1.1"
+    "nyc": "^14.1.1",
+    "pg": "^7.12.1"
   },
   "devDependencies": {
     "@babel/plugin-transform-modules-commonjs": "^7.6.0",
     "babel-plugin-istanbul": "^5.2.0",
     "chai": "^4.2.0",
     "chai-http": "^4.3.0",
-    "coveralls": "^3.0.6",
     "eslint": "^6.2.1",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-plugin-import": "^2.18.2",
