+++ server/v1/controllers/entry.js
