+++ server/v1/middleware/checkAuth.js
