+++ server/v1/middleware/checkEntryAll.js
