+++ server/v1/middleware/checkUserEntry.js
